Add minor changes
